moving ApiSource enum to its own class
